refactor(fused_marlin_moe): merge W4A16 kernel into one file, add deepseek-v4-flash shapes
