+++ benchmark/test_fused_marlin_moe.py
@@ -177,6 +177,15 @@ class FusedMarlinMoEBenchmark(base.Benchmark):
             (64, 512, 4096, 1024, 10),
             (128, 512, 4096, 1024, 10),
             (256, 512, 4096, 1024, 10),
+            # DeepSeek-V4-Flash
+            (1, 256, 4096, 2048, 6),
+            (4, 256, 4096, 2048, 6),
+            (8, 256, 4096, 2048, 6),
+            (16, 256, 4096, 2048, 6),
+            (32, 256, 4096, 2048, 6),
+            (64, 256, 4096, 2048, 6),
+            (128, 256, 4096, 2048, 6),
+            (256, 256, 4096, 2048, 6),
         ]
 
     def get_input_iter(self, cur_dtype):
